+++ src/snmpmsg.cpp
@@ -149,7 +149,7 @@ int convertVbToSmival(const Vb& tempvb, SmiVALUE* smival)
         if (smival->value.oid.len > 0)
         {
             smival->value.oid.ptr =
-                (SmiLPUINT32) new unsigned long[smival->value.oid.len];
+                (SmiLPUINT32) new uint32_t[smival->value.oid.len];
             if (smival->value.oid.ptr)
             {
                 for (int i = 0; i < (int)smival->value.oid.len; i++)
@@ -227,8 +227,8 @@ int SnmpMessage::load(
 #ifdef _SNMPv3
     raw_pdu->msgid = pdu->get_message_id();
 #endif
-    raw_pdu->errstat  = (unsigned long)pdu->get_error_status();
-    raw_pdu->errindex = (unsigned long)pdu->get_error_index();
+    raw_pdu->errstat  = (uint32_t)pdu->get_error_status();
+    raw_pdu->errindex = (uint32_t)pdu->get_error_index();
 
     // if its a V1 trap then load up other values
     // for v2, use normal pdu format
@@ -349,7 +349,7 @@ int SnmpMessage::load(
         // timestamp
         TimeTicks timestamp;
         pdu->get_notify_timestamp(timestamp);
-        raw_pdu->time = (unsigned long)timestamp;
+        raw_pdu->time = (uint32_t)timestamp;
     }
 
     // if its a v2 trap then we need to make a few adjustments
@@ -518,7 +518,7 @@ int SnmpMessage::load(
 }
 
 // load up a SnmpMessage
-int SnmpMessage::load(unsigned char* data, unsigned long len)
+int SnmpMessage::load(unsigned char* data, uint32_t len)
 {
     bufflen    = MAX_SNMP_PACKET;
     valid_flag = false;
@@ -678,13 +678,13 @@ int SnmpMessage::unload(Pdu& pdu,           // Pdu object
             // octet string
         case sNMP_SYNTAX_OCTETS: {
             OctetStr octets(
-                (unsigned char*)vp->val.string, (unsigned long)vp->val_len);
+                (unsigned char*)vp->val.string, (uint32_t)vp->val_len);
             tempvb.set_value(octets);
         }
         break;
         case sNMP_SYNTAX_OPAQUE: {
             OpaqueStr octets(
-                (unsigned char*)vp->val.string, (unsigned long)vp->val_len);
+                (unsigned char*)vp->val.string, (uint32_t)vp->val_len);
             tempvb.set_value(octets);
         }
         break;
@@ -707,7 +707,7 @@ int SnmpMessage::unload(Pdu& pdu,           // Pdu object
 
             // timeticks
         case sNMP_SYNTAX_TIMETICKS: {
-            TimeTicks timeticks((unsigned long)*(vp->val.integer));
+            TimeTicks timeticks((uint32_t) * (vp->val.integer));
             tempvb.set_value(timeticks);
             if ((vb_nr == 1)
                 && ((raw_pdu->command == sNMP_PDU_TRAP)
@@ -723,14 +723,14 @@ int SnmpMessage::unload(Pdu& pdu,           // Pdu object
 
             // 32 bit counter
         case sNMP_SYNTAX_CNTR32: {
-            Counter32 counter32((unsigned long)*(vp->val.integer));
+            Counter32 counter32((uint32_t) * (vp->val.integer));
             tempvb.set_value(counter32);
         }
         break;
 
             // 32 bit gauge
         case sNMP_SYNTAX_GAUGE32: {
-            Gauge32 gauge32((unsigned long)*(vp->val.integer));
+            Gauge32 gauge32((uint32_t) * (vp->val.integer));
             tempvb.set_value(gauge32);
         }
         break;
@@ -770,7 +770,7 @@ int SnmpMessage::unload(Pdu& pdu,           // Pdu object
             /* Not distinguishable from Gauge32
                 case sNMP_SYNTAX_UINT32:
                   {
-                    SnmpUInt32 uint32((unsigned long) *(vp->val.integer));
+                    SnmpUInt32 uint32((uint32_t) *(vp->val.integer));
                     tempvb.set_value(uint32);
                   }
                   break;
